demo updates

+++ app/Http/Controllers/Web/Admin/VehicleTypeController.php
@@ -103,6 +103,11 @@ class VehicleTypeController extends BaseController
      */
     public function store(CreateVehicleTypeRequest $request)
     {
+         if (env('APP_FOR')=='demo') {
+            $message = trans('succes_messages.you_are_not_authorised');
+
+            return redirect('types')->with('warning', $message);
+        }
         // dd($request->transport_type);
         $created_params = $request->only(['name', 'capacity','is_accept_share_ride','description','supported_vehicles','short_description', 'transport_type']);
 
@@ -153,6 +158,11 @@ class VehicleTypeController extends BaseController
      */
     public function update(UpdateVehicleTypeRequest $request, VehicleType $vehicle_type)
     {
+        if (env('APP_FOR')=='demo') {
+            $message = trans('succes_messages.you_are_not_authorised');
+
+            return redirect('types')->with('warning', $message);
+        }
         // dd($request->all());
         $this->validateAdmin();
         $created_params = $request->only(['name', 'capacity','is_accept_share_ride','description','supported_vehicles','short_description','transport_type']);
@@ -174,6 +184,12 @@ class VehicleTypeController extends BaseController
     }
     public function toggleStatus(VehicleType $vehicle_type)
     {
+        if (env('APP_FOR')=='demo') {
+            $message = trans('succes_messages.you_are_not_authorised');
+
+            return redirect('types')->with('warning', $message);
+        }
+        
         $status = $vehicle_type->active == 1 ? 0 : 1;
         $vehicle_type->update([
             'active' => $status
@@ -196,6 +212,12 @@ class VehicleTypeController extends BaseController
 
     public function delete(VehicleType $vehicle_type)
     {
+        if (env('APP_FOR')=='demo') {
+            $message = trans('succes_messages.you_are_not_authorised');
+
+            return redirect('types')->with('warning', $message);
+        }
+
         $vehicle_type->delete();
 
         $message = trans('succes_messages.vehicle_type_deleted_succesfully');
